+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/types/CLIPACTIONS.java
@@ -12,14 +12,18 @@
  * Lesser General Public License for more details.
  * 
  * You should have received a copy of the GNU Lesser General Public
- * License along with this library. */
+ * License along with this library.
+ */
 package com.jpexs.decompiler.flash.types;
 
 import com.jpexs.decompiler.flash.types.annotations.Reserved;
 import com.jpexs.decompiler.flash.types.annotations.SWFType;
 import java.io.Serializable;
+import java.lang.reflect.Field;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 /**
  * Define event handlers for a sprite character
@@ -41,4 +45,24 @@ public class CLIPACTIONS implements Serializable {
      * Individual event handlers
      */
     public List<CLIPACTIONRECORD> clipActionRecords = new ArrayList<>();
+
+    public void calculateAllEventFlags() {
+
+        Field[] fields = allEventFlags.getClass().getDeclaredFields();
+        try {
+            for (Field f : fields) {
+                if (!f.getName().startsWith("clipEvent")) {
+                    continue;
+                }
+                f.set(allEventFlags, false);
+                for (CLIPACTIONRECORD car : clipActionRecords) {
+                    if (f.getBoolean(car.eventFlags)) {
+                        f.set(allEventFlags, true);
+                    }
+                }
+            }
+        } catch (IllegalArgumentException | IllegalAccessException ex) {
+            Logger.getLogger(CLIPACTIONS.class.getName()).log(Level.SEVERE, null, ex);
+        }
+    }
 }